Read files stored in an OneDrive

This function reads files stored in an OneDrive

Usage

OneDrive_file(file, .function, ...)

Arguments

file

File path

.function

Function that will be used to reading the file

...

.function args.

Examples

 OneDrive_file("path/file.csv", read.csv)
#> httpuv not installed, defaulting to device code authentication
#> Using device_code flow
#> Creating Microsoft Graph login for default tenant
#> httpuv not installed, defaulting to device code authentication
#> Using device_code flow
#> To sign in, use a web browser to open the page https://microsoft.com/devicelogin and enter the code QSKHJ8LXN to authenticate. 
#> Waiting for device code in browser...
#> Press Esc/Ctrl + C to abort
#> Error in process_aad_response(res): Bad Request (HTTP 400). Failed to obtain Azure Active Directory token. Message:
#> AADSTS70020: The provided value for the input parameter 'device_code' is not valid. This device code has expired.
#> Trace ID: de628027-bff8-4177-a83e-5955b2e13400
#> Correlation ID: 980e6be1-4ec1-4158-85a7-8d3e2788d14e
#> Timestamp: 2022-10-09 16:14:12Z.